feat: move sign coordinator logic into its own thread
`SignerDBListener` struct is for a new thread that is always processing
StackerDB messages from the signers during a mining tenure.

`SignerCoordinator` is the interface that the miner uses with the
`SignerDBListener`, to propose a block and wait for signatures.
